Create subscription after completing onboarding

+++ app/forms/onboarding/how_much_should_an_individual_give.rb
@@ -5,7 +5,7 @@ module Onboarding
     message "First up, let's think about the obligations that we have as individuals in our society."
     message 'As a percentage of pre-tax income, how much do you believe an individual should give to charity?'
 
-    response_type :float
+    response_type :decimal
 
     allowed_response 0.005, '0.5%'
     allowed_response 0.01, '1%'
